+++ mod-source-record-storage-server/src/test/java/org/folio/dao/util/EdifactUtilTest.java
@@ -1,27 +1,26 @@
 package org.folio.dao.util;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-
-import java.io.File;
-import java.io.IOException;
-
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-
+import io.xlate.edi.stream.EDIStreamException;
 import org.folio.rest.jaxrs.model.SourceRecord;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.BlockJUnit4ClassRunner;
 
-import io.xlate.edi.stream.EDIStreamException;
+import java.io.File;
+import java.io.IOException;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 @RunWith(BlockJUnit4ClassRunner.class)
 public class EdifactUtilTest {
 
   private static final String SOURCE_RECORD_PATH = "src/test/resources/mock/sourceRecords/e4cfe577-4015-46d8-a54d-7c9b34796955.json";
+  private static final String SOURCE_RECORD_WITH_IGNORED_CODES_PATH = "src/test/resources/mock/sourceRecords/4ca9d8ac-9de5-432a-83ee-15832f09e868.json";
 
   private static final String INVALID_RAW_EDIFACT_RECORD =
     "UNA:+.? '" +
@@ -67,6 +66,15 @@ public class EdifactUtilTest {
 
   @Test
   public void shouldFormatEdifactAlreadyFormatted() throws IOException, EDIStreamException {
+    String rawEdifact = sourceRecord.getParsedRecord().getFormattedContent();
+    String formattedEdifact = EdifactUtil.formatEdifact(rawEdifact);
+    assertNotNull(formattedEdifact);
+    assertEquals(sourceRecord.getParsedRecord().getFormattedContent(), formattedEdifact);
+  }
+
+  @Test
+  public void shouldFormatEdifactWithIgnoredCodeValues() throws IOException, EDIStreamException {
+    SourceRecord sourceRecord = new ObjectMapper().readValue(new File(SOURCE_RECORD_WITH_IGNORED_CODES_PATH), SourceRecord.class);
     String rawEdifact = sourceRecord.getParsedRecord().getFormattedContent();
     String formattedEdifact = EdifactUtil.formatEdifact(rawEdifact);
     assertNotNull(formattedEdifact);